backend/pyproject.toml
[project]
name = "behavioral-predictor"
version = "1.0.0"
description = "Production-grade browser-deployable behavioral prediction system"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Caching & Queue
    "redis>=5.0.0",
    "celery[redis]>=5.3.0",
    # ML
    "torch>=2.2.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.17.0",
    "onnxscript>=0.1.0",
    "scikit-learn>=1.4.0",
    # Data
    "pyarrow>=15.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    # Validation
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    # Observability
    "structlog>=24.1.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "prometheus-client>=0.20.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
